# Performance improvements wrt g_MemMap