--- a/ReClass/ReClassEx.rc
+++ b/ReClass/ReClassEx.rc
@@ -1,4 +1,5 @@
-﻿// Microsoft Visual C++ generated resource script.
+#pragma code_page(65001)
+// Microsoft Visual C++ generated resource script.
 //
 #include "resource.h"