… TestPlugin/plugins/TestPlugin_dbg.lib
… TestPlugin/plugins/TestPlugin_dbg.rc-plugin64